Refactor: refactor header into seperate component
